Add implementations and tests of Stirling numbers in terms of SymEngine types.
Also minor edits and clean-ups.

+++ cpp/cache.hpp
@@ -44,5 +44,5 @@ public:
 	}
 
 private:
-	std::unordered_map<int, R> cache;
+	std::unordered_map<long, R> cache;
 };
